requirements.txt
asyncio
python-kasa